+++ tests/DataSet/DataSetTest.php
@@ -6,14 +6,14 @@ namespace Yiisoft\Validator\Tests\DataSet;
 
 use Yiisoft\Validator\Result;
 use Yiisoft\Validator\Tests\Stub\DataSet;
-use Yiisoft\Validator\Tests\Stub\FakeValidatorFactory;
+use Yiisoft\Validator\Tests\Support\ValidatorFactory;
 
 final class DataSetTest extends AbstractDataSetTest
 {
     protected function validate(array $dataSet, array $rules): Result
     {
         $dataObject = new DataSet($dataSet);
-        $validator = FakeValidatorFactory::make();
+        $validator = ValidatorFactory::make();
 
         return $validator->validate($dataObject, $rules);
     }
