+++ tests/Support/ValidatorFactory.php
@@ -2,12 +2,13 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Validator\Tests\Stub;
+namespace Yiisoft\Validator\Tests\Support;
 
 use Yiisoft\Validator\SimpleRuleHandlerContainer;
+use Yiisoft\Validator\Tests\Stub\TranslatorFactory;
 use Yiisoft\Validator\Validator;
 
-final class FakeValidatorFactory
+final class ValidatorFactory
 {
     public static function make(): Validator
     {
