+++ tests/ValidatorTest.php
@@ -23,7 +23,7 @@ use Yiisoft\Validator\RuleInterface;
 use Yiisoft\Validator\SimpleRuleHandlerContainer;
 use Yiisoft\Validator\SkipOnEmptyCallback\SkipOnNull;
 use Yiisoft\Validator\Tests\Stub\DataSet;
-use Yiisoft\Validator\Tests\Stub\FakeValidatorFactory;
+use Yiisoft\Validator\Tests\Support\ValidatorFactory;
 use Yiisoft\Validator\Tests\Stub\NotNullRule\NotNull;
 use Yiisoft\Validator\Tests\Stub\ObjectWithDataSet;
 use Yiisoft\Validator\Tests\Stub\Rule;
@@ -37,7 +37,7 @@ class ValidatorTest extends TestCase
     public function testAddingRulesViaConstructor(): void
     {
         $dataObject = new DataSet(['bool' => true, 'int' => 41]);
-        $validator = FakeValidatorFactory::make();
+        $validator = ValidatorFactory::make();
         $result = $validator->validate($dataObject, [
             'bool' => [new Boolean()],
             'int' => [
@@ -78,7 +78,7 @@ class ValidatorTest extends TestCase
      */
     public function testDiverseTypes($dataSet): void
     {
-        $validator = FakeValidatorFactory::make();
+        $validator = ValidatorFactory::make();
         $result = $validator->validate($dataSet, [new Required()]);
 
         $this->assertTrue($result->isValid());
@@ -86,7 +86,7 @@ class ValidatorTest extends TestCase
 
     public function testNullAsDataSet(): void
     {
-        $validator = FakeValidatorFactory::make();
+        $validator = ValidatorFactory::make();
         $result = $validator->validate(null, ['property' => [new CompareTo(null)]]);
 
         $this->assertTrue($result->isValid());
@@ -94,7 +94,7 @@ class ValidatorTest extends TestCase
 
     public function testPreValidation(): void
     {
-        $validator = FakeValidatorFactory::make();
+        $validator = ValidatorFactory::make();
         $result = $validator->validate(
             new DataSet(['property' => '']),
             ['property' => [new Required(when: static fn (mixed $value, ?ValidationContext $context): bool => false)]],
@@ -107,7 +107,7 @@ class ValidatorTest extends TestCase
     {
         $ruleHandler = new class () {
         };
-        $validator = FakeValidatorFactory::make();
+        $validator = ValidatorFactory::make();
 
         $this->expectException(RuleHandlerInterfaceNotImplementedException::class);
         $validator->validate(new DataSet(['property' => '']), [
@@ -135,7 +135,7 @@ class ValidatorTest extends TestCase
     {
         $this->expectException(RuleHandlerNotFoundException::class);
 
-        $validator = FakeValidatorFactory::make();
+        $validator = ValidatorFactory::make();
         $validator->validate(new DataSet(['property' => '']), [
             'property' => [
                 new class () implements RuleInterface {
@@ -302,7 +302,7 @@ class ValidatorTest extends TestCase
      */
     public function testRequired(array|null $rules, DataSetInterface $dataSet, array $expectedErrors): void
     {
-        $validator = FakeValidatorFactory::make();
+        $validator = ValidatorFactory::make();
         $result = $validator->validate($dataSet, $rules);
         $this->assertEquals($expectedErrors, $result->getErrors());
     }
@@ -310,7 +310,7 @@ class ValidatorTest extends TestCase
     public function skipOnEmptyDataProvider(): array
     {
         $translator = (new TranslatorFactory())->create();
-        $validator = FakeValidatorFactory::make();
+        $validator = ValidatorFactory::make();
         $rules = [
             'name' => [new HasLength(min: 8)],
             'age' => [new Number(asInteger: true, min: 18)],
@@ -759,7 +759,7 @@ class ValidatorTest extends TestCase
     {
         $object = new ObjectWithAttributesOnly();
 
-        $validator = FakeValidatorFactory::make();
+        $validator = ValidatorFactory::make();
 
         $result = $validator->validate($object);
